Rename SQLite connection accessor
